% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_talk.R
\name{extract_body_paragraphs_df}
\alias{extract_body_paragraphs_df}
\title{Produce paragraphs}
\usage{
extract_body_paragraphs_df(rv_doc)
}
\arguments{
\item{rv_doc}{rvest::read_html() document}
}
\value{
dataframe with paragraphs
}
\description{
Produce paragraphs
}
