% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_talk.R
\name{extract_element}
\alias{extract_element}
\title{Extract html document elements}
\usage{
extract_element(html_document, element)
}
\arguments{
\item{html_document}{rvest::read_html() document}

\item{element}{class you want to extract (use Selector Gadget)}
}
\value{
dataframe
}
\description{
rvest::read_html("https://www.churchofjesuschrist.org/study/general-conference/1971/04/kingdom-of-god?lang=eng") %>% extract_element("#title1")
}
