% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_talk.R
\name{extract_metadata}
\alias{extract_metadata}
\title{Extract title, author, and kicker from a url and return as a row in a
dataframe.}
\usage{
extract_metadata(html_document, url)
}
\description{
Extract title, author, and kicker from a url and return as a row in a
dataframe.
}
