% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_talk_urls.R
\name{extract_session_hrefs}
\alias{extract_session_hrefs}
\title{Extract Session hrefs}
\usage{
extract_session_hrefs(html_docmap, session_id)
}
\arguments{
\item{html_docmap}{An rvest docmap scrape from
scrape_conference_html_doc_map()}

\item{session_id}{Integer for session you want to extract}
}
\value{
hrefs for the session, which includes the Session href in addition
to the talk refs.
}
\description{
Extract Session hrefs
}
\examples{
scrape_conference_html_doc_map(2019, 4) \%>\%
  extract_session_hrefs(session_id = 1) \%>\%
  parse_session_urls()
}
