% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_talk.R
\name{extract_url_from_rv_doc}
\alias{extract_url_from_rv_doc}
\title{Extract url from rv_doc}
\usage{
extract_url_from_rv_doc(rv_doc)
}
\arguments{
\item{rv_doc}{rvest::read_html() object}
}
\value{
string
}
\description{
Extract url from rv_doc
}
