% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_talk_urls.R
\name{parse_path_for_name}
\alias{parse_path_for_name}
\title{Parse path for name}
\usage{
parse_path_for_name(path)
}
\arguments{
\item{path}{file path with no extension}
}
\value{
Upper Case String
}
\description{
Parse path for name
}
\examples{
parse_path_for_name(path = "/path/to/hello-world-")
}
