% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_talk_urls.R
\name{parse_session_urls}
\alias{parse_session_urls}
\title{Parse Session URLs
Take a vector of session hrefs, use the first value
as the session name, then the other values as the
session talks}
\usage{
parse_session_urls(session_hrefs)
}
\arguments{
\item{session_hrefs}{vector}
}
\value{
nested tibble
}
\description{
Parse Session URLs
Take a vector of session hrefs, use the first value
as the session name, then the other values as the
session talks
}
