% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_talk.R
\name{parse_url}
\alias{parse_url}
\title{Scrape an individual General Conference URL}
\usage{
parse_url(url)
}
\arguments{
\item{url}{raw url}
}
\value{
string
}
\description{
Helps scrape a talk
}
\details{
See unit tests for edge case urls.
Simple string extractor. Removes ?lang=eng and other stuff
}
