% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_talk_urls.R
\name{scrape_conference_html_doc_map}
\alias{scrape_conference_html_doc_map}
\title{Scrape HTML doc map from Conference URL}
\usage{
scrape_conference_html_doc_map(year, month)
}
\arguments{
\item{year}{Year (integer)}

\item{month}{Month (integer)}
}
\value{
Rvest object
}
\description{
Given a year and a month, pull the entire .doc-map class
object from the Conference URL. This will be parsed
by downstream objects
}
\examples{
scrape_conference_html_doc_map(2017, 4)
scrape_conference_html_doc_map(1971, 10)
scrape_conference_html_doc_map(1985, 10)
}
