% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_session.R
\name{scrape_conference_talks}
\alias{scrape_conference_talks}
\title{Scrapes all conference talks for a sessions}
\usage{
scrape_conference_talks(year, month, path, loop_method = 1)
}
\arguments{
\item{year}{Year}

\item{month}{Month}
}
\value{
Writes out session to /data/sessions/<year><month>.rds
}
\description{
For one-off sessions or debugging, see new-sessions.Rmd.
}
