% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_talk_urls.R
\name{scrape_conference_urls}
\alias{scrape_conference_urls}
\title{Main function to scrape all conference talk urls
For a given year-month conference, return a nested tibble of all sessions
with a tibble-column containing the dataframes}
\usage{
scrape_conference_urls(year, month)
}
\arguments{
\item{year}{year}

\item{month}{month}
}
\value{
tibble
}
\description{
Main function to scrape all conference talk urls
For a given year-month conference, return a nested tibble of all sessions
with a tibble-column containing the dataframes
}
\examples{
scrape_conference_urls(2019, 10)
scrape_conference_urls(1971, 4)
}
