% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_talk.R
\name{scrape_talk}
\alias{scrape_talk}
\title{Scrape general conference talk}
\usage{
scrape_talk(url)
}
\arguments{
\item{url}{general conference https}
}
\value{
dataframe
}
\description{
Scrape general conference talk
}
